requirements.txt
aiohttp~=3.9.0
Brotli~=1.1.0
yarl~=1.9.3
gunicorn==19.9.0
lxml==4.9.2
redis==3.5.3

beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
google==3.0.0
idna==3.4
protobuf==4.24.3
protobuf-inspector==0.2
requests==2.31.0
soupsieve==2.5
urllib3==2.0.4
websocket-client==1.6.3

APScheduler==3.10.0;python_version>="3.10"
APScheduler==3.2.0;python_version<"3.10"
click==8.0.1;python_version>"3.6"
click==7.0;python_version<="3.6"
Flask==2.1.1;python_version>"3.6"
Flask==1.0;python_version<="3.6"
werkzeug==2.1.0;python_version>"3.6"
werkzeug==0.15.5;python_version<="3.6"

pandas==2.2.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
kazoo==2.10.0
dingtalkchatbot==1.5.7
aiohttp_proxy==0.1.2
python-redis-lock==4.0.0
tzlocal==2.1

jieba==0.42.1